Integrated hbm2ddl.auto create
Reconfigured the table classes definitions and the database class and the controller class in order to work with the property “hbm2ddl.auto create“. All the types of the parameters that reference policies id were changed to Integer
ClearUsers() and ClearPolicies() were removed from the database class. The tables are now reset upon program launch
Added dialect property to hibernate.cfg.xml

+++ src/com/corso/policysystem/Policy.java
@@ -10,7 +10,8 @@ public class Policy {
     @Id
     @GeneratedValue
     @Column(name = "id")
-    private Long id;
+    private Integer id;
+    //Long type creates a BIGINT id and it's way too big to be used as id
 
     @Column(name = "description")
     private String description;
@@ -31,7 +32,7 @@ public class Policy {
         this.cost = cost;
     }
 
-    public Long getId(){
+    public Integer getId(){
         return id;
     }
